requirements.txt
flask==2.2.3
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
langchain-openai==0.0.2
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
flask-uploads==0.2.1
pdfplumber==0.10.0
chromadb==0.4.15
